feat: replication testing (#15)
The replication tests use the output of the upload tests (specifically
`shards.csv`) to request replications for successfully uploaded shards.
The script simply makes a `space/blob/replicate` invocation to the
upload service and waits for the transfer tasks to complete (or
fail/timeout) by polling the upload service receipts endpoint.

Since the replication tests are performed after the upload tests the
location commitment needed for the `space/blob/replicate` invocation
needs to be retrieved from the storage node that issued it. We
indirectly do this by querying the indexing service.

---------

Co-authored-by: Vicente Olmedo <vicente@storacha.network>